+++ forui/lib/widgets/select.dart
@@ -2,5 +2,3 @@
 ///
 /// TODO: This is a placeholder for the select widget.
 library forui.widgets.select;
-
-export '../src/widgets/select/select_controller.dart';