+++ samples/lib/widgets/popover_menu.dart
@@ -15,7 +15,6 @@ class PopoverMenuPage extends StatefulSample {
 
 class _State extends StatefulSampleState<PopoverMenuPage> with SingleTickerProviderStateMixin {
   late final controller = FPopoverController(vsync: this);
-  late final selectController = FSelectController<String>.radio();
 
   @override
   Widget sample(BuildContext context) => FHeader(
@@ -48,7 +47,6 @@ class _State extends StatefulSampleState<PopoverMenuPage> with SingleTickerProvi
   @override
   void dispose() {
     controller.dispose();
-    selectController.dispose();
     super.dispose();
   }
 }