+++ samples/lib/widgets/tile_group.dart
@@ -116,7 +116,7 @@ class _MergeTileGroup extends StatefulWidget {
 }
 
 class _MergeTileGroupState extends State<_MergeTileGroup> {
-  late final controller = FSelectController.radio(value: 'List');
+  late final controller = FMultiValueNotifier.radio(value: 'List');
 
   @override
   Widget build(BuildContext context) => FTileGroup.merge(