+++ forui_hooks/CHANGELOG.md
@@ -1,7 +1,15 @@
 ## 0.11.0
-* Rename `useFRadioSelectGroupController` to `useFRadioSelectController`.
-* Rename `useFMultiSelectGroupController` to `useFSelectController`.
+* Add `useFSelectGroupController`.
+* Add `useFRadioSelectGroupController`.
 
+* Add `useFSelectTileGroupController`.
+* Add `useFRadioSelectTileGroupController`.
+
+* Add `useFSelectMenuTileController`.
+* Add `useFRadioSelectMenuTileController`.
+
+* Rename `useFMultiSelectGroupController` to `useFMultiValueNotifier`.
+* Rename `useFRadioSelectGroupController` to `useFRadioMultiValueNotifier`.
 
 ## 0.10.0
 Bump `flutter_hooks` version to `0.21.0`.